+++ examples/src/samples/RangeSwitcher/RangeSwitcher.tsx
@@ -1,5 +1,10 @@
 import { Button, ButtonGroup } from "@mui/material";
-import { AreaSeries, Chart } from "lightweight-charts-react-components";
+import {
+  AreaSeries,
+  Chart,
+  TimeScale,
+  TimeScaleFitContentTrigger,
+} from "lightweight-charts-react-components";
 import { colors } from "@/colors";
 import { withChartCommonOptions } from "@/common/chartCommonOptions";
 import { typedObjectKeys } from "@/common/utils";
@@ -57,6 +62,9 @@ const RangeSwitcher = () => {
         containerProps={{ style: { flexGrow: "1" } }}
       >
         <AreaSeries options={seriesCustomOptions} data={data} />
+        <TimeScale>
+          <TimeScaleFitContentTrigger deps={[]} />
+        </TimeScale>
       </Chart>
     </ChartWidgetCard>
   );