+++ lib/src/scales/index.ts
@@ -1,8 +1,10 @@
 export { TimeScale } from "./TimeScale";
 export { PriceScale } from "./PriceScale";
+export { TimeScaleFitContentTrigger } from "./TimeScaleFitContentTrigger";
 export type {
   PriceScaleProps,
   TimeScaleProps,
   PriceScaleApiRef,
   TimeScaleApiRef,
+  TimeScaleFitContentTriggerProps,
 } from "./types";